==> 1.create_mapaffil.ipynb <==
# Create mapaffil2018.csv.gz from original download

Download `mapaffil2018.tsv.gz` from https://databank.illinois.edu/datasets/IDB-2556310 and save to "data" folder

import pandas as pd
import csv
import gzip

df = pd.read_csv('data/mapaffil2018.tsv.gz', compression='gzip', delimiter="\t", on_bad_lines='skip')

df.to_csv("data/mapaffil2018.csv.gz", index=False)
