==> 2.create_unique_sorted_mapaffil.ipynb <==
# Reduce MapAffil to unique affiliation strings sorted by occurence

import pandas as pd

mapaffil = pd.read_csv("data/mapaffil2018.csv.gz", on_bad_lines='skip') 

# Get the count of each affiliation and sort in descending order
mapaffil_counts = mapaffil["affiliation"].value_counts().sort_values(ascending=False)

# Create a DataFrame with affiliation and count columns
sorted_mapaffil = pd.DataFrame({"affiliation": mapaffil_counts.index, "count": mapaffil_counts})

# Join the sorted DataFrame with the original mapaffil DataFrame on affiliation column
sorted_mapaffil = sorted_mapaffil.merge(mapaffil, on="affiliation")

Reduce Mapaffil to relevant columns and remove any duplicates

sorted_mapaffil = sorted_mapaffil[['affiliation', 'city', 'state', 'country', 'lat', 'lon', 'count']]

unique_sorted_mapaffil = sorted_mapaffil.drop_duplicates() # 20,326,273 unique affiliations out of 52,915,684 total authorships

unique_sorted_mapaffil.to_csv("data/mapaffil2018-unique_sorted.csv.gz", index=False)
