==> 4.save_model.ipynb <==
# Create and save classification model

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn import svm
import joblib

clean_spacy_mapaffil = pd.read_parquet("data/clean_spacy_mapaffil.parquet", engine="fastparquet") 

clean_spacy_mapaffil_subset = clean_spacy_mapaffil.head(1000000)

# Assign training datatsets
X_train =  list(clean_spacy_mapaffil_subset['affiliation']) 
y_train = list(clean_spacy_mapaffil_subset['city']) 

vectorizer = TfidfVectorizer(stop_words="english", decode_error="ignore")
vectorizer.fit(X_train)
features_train = vectorizer.transform(X_train)

vectorizer_filename = 'geoinference_vectorizer_1mil.joblib.lzma'
joblib.dump(vectorizer, vectorizer_filename)

clf = svm.LinearSVC(dual=True)
clf.fit(features_train, y_train)

# 6.5 hours

model_filename = 'geoinference_linearsvc_1mil.joblib.lzma'
joblib.dump(clf, model_filename)
